Yolov5 Python Azure Function Blob Triggered/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.12.0
onnxruntime==1.13.1
requests==2.28.1
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.7.1+cpu 
torchvision==0.8.2+cpu
azure.storage.blob==12.14.1
pathlib==1.0.1

# YOLOv5 requirements
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
#matplotlib>=3.2.2
numpy==1.24.1
opencv-python==4.7.0.68
Pillow==9.4.0
PyYAML==6.0
scipy==1.10.0
tqdm==4.64.1
# protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012

# Logging -------------------------------------
#tensorboard>=2.4.1
# wandb
# clearml

# Plotting ------------------------------------
pandas==1.5.2

psutil  # system utilization
thop>=0.1.1  # FLOPs computation